+++ wcomponents-core/src/main/java/com/github/bordertech/wcomponents/container/TemplateRenderInterceptor.java
@@ -1,20 +1,17 @@
 package com.github.bordertech.wcomponents.container;
 
 import com.github.bordertech.wcomponents.RenderContext;
-import com.github.bordertech.wcomponents.Request;
 import com.github.bordertech.wcomponents.UIContext;
 import com.github.bordertech.wcomponents.UIContextHolder;
-import com.github.bordertech.wcomponents.servlet.ServletRequest;
 import com.github.bordertech.wcomponents.servlet.WebXmlRenderContext;
+import com.github.bordertech.wcomponents.template.HandlebarsRendererImpl;
 import com.github.bordertech.wcomponents.template.TemplateRenderer;
 import com.github.bordertech.wcomponents.template.TemplateRendererFactory;
-import com.github.bordertech.wcomponents.util.ConfigurationProperties;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
-import javax.servlet.http.HttpServletRequest;
 
 /**
  * Render templates on the server (i.e. Handlebars).
@@ -26,34 +23,15 @@ public class TemplateRenderInterceptor extends InterceptorComponent {
 	 */
 	private static final TemplateRenderer TEMPLATE_RENDERER = TemplateRendererFactory.newInstance(TemplateRendererFactory.TemplateEngine.HANDLEBARS);
 
-	private static final Map<String, Object> CONTEXT;
-
-	private boolean doRender = false;
+	/**
+	 * No CACHE option.
+	 */
+	private static final Map<String, Object> OPTIONS;
 
 	static {
-		Map<String, Object> tmpContext = new HashMap<>();
-		tmpContext.put("rendered", "server");
-		CONTEXT = Collections.unmodifiableMap(tmpContext);
-	}
-
-	/**
-	 * Override preparePaint in order to perform processing specific to this interceptor.
-	 *
-	 * @param request the request being responded to.
-	 */
-	@Override
-	public void preparePaint(final Request request) {
-		// TODO this should check the the response is actually HTML / Handlebars (not XML)
-		String mode = ConfigurationProperties.getTemplateRenderingMode();
-		doRender = "on".equals(mode);
-		if (!doRender && "sniff".equals(mode) && request instanceof ServletRequest) {
-			HttpServletRequest httpServletRequest = ((ServletRequest) request).getBackingRequest();
-			String userAgentString = httpServletRequest.getHeader("User-Agent");
-			if (userAgentString.indexOf("MSIE 8.0") > 0 || userAgentString.indexOf("MSIE 9.0;") > 0) {
-				doRender = true;
-			}
-		}
-		super.preparePaint(request);
+		HashMap<String, Object> map = new HashMap<>();
+		map.put(HandlebarsRendererImpl.USE_CACHE, "false");
+		OPTIONS = Collections.unmodifiableMap(map);
 	}
 
 	/**
@@ -61,12 +39,6 @@ public class TemplateRenderInterceptor extends InterceptorComponent {
 	 */
 	@Override
 	public void paint(final RenderContext renderContext) {
-
-		if (!doRender) {
-			super.paint(renderContext);
-			return;
-		}
-
 		UIContext uic = UIContextHolder.getCurrent();
 
 		// Set up a render CONTEXT to buffer the Payload.
@@ -80,6 +52,6 @@ public class TemplateRenderInterceptor extends InterceptorComponent {
 		PrintWriter writer = webRenderContext.getWriter();
 
 		// Transform handlebars
-		TEMPLATE_RENDERER.renderInline(outputBuffer.toString(), CONTEXT, Collections.EMPTY_MAP, writer, Collections.EMPTY_MAP);
+		TEMPLATE_RENDERER.renderInline(outputBuffer.toString(), Collections.EMPTY_MAP, Collections.EMPTY_MAP, writer, OPTIONS);
 	}
 }