+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/render/webxml/WMenuRenderer_Test.java
@@ -109,4 +109,18 @@ public class WMenuRenderer_Test extends AbstractWebXmlRendererTestCase {
 		assertXpathEvaluatesTo("xl", "//ui:menu/ui:margin/@west", menu);
 	}
 
+	@Test
+	public void testXssEscaping() throws IOException, SAXException, XpathException {
+		WMenu menu = new WMenu();
+		menu.add(new WMenuItem("test"));
+		
+		assertSafeContent(menu);
+
+		menu.setToolTip(getMaliciousAttribute());
+		assertSafeContent(menu);
+
+		menu.setAccessibleText(getMaliciousAttribute());
+		assertSafeContent(menu);
+	}
+
 }